shopfloor_mobile_base: reorganize registries
