+++ shopfloor_mobile_base/static/wms/src/services/color_registry.js
@@ -4,7 +4,30 @@
  * License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
  */
 
-import {ColorRegistry} from "./registry.js";
+export class ColorRegistry {
+    constructor(theme, _default = "light") {
+        this.themes = {};
+        this.default_theme = _default;
+    }
+
+    add_theme(colors, theme) {
+        if (_.isUndefined(theme)) theme = this.default_theme;
+        this.themes[theme] = colors;
+    }
+
+    color_for(key, theme) {
+        if (_.isUndefined(theme)) theme = this.default_theme;
+        if (!this.themes[theme]) {
+            console.log("Theme", theme, "not registered.");
+            return null;
+        }
+        return this.themes[theme][key];
+    }
+
+    get_themes() {
+        return this.themes;
+    }
+}
 
 export var color_registry = new ColorRegistry();
